Fix clean up logic to cancel fetch request if another is made
This was causing an issue where EditPost was receiving incorrect data and
passing incorrect props to PostForm. This caused the post to be undefined
when state was initialized the first time, thus it stayed undefined
for subsequent attempts to access the component. The reason the PostForm
component was reached is because for the first fetch, the isLoading and
fetch error were truthy despite the flawed state so the if conditions
were bypassed and the component was rendered with flawed props.